+++ packages/elements/checkbox/src/CheckboxGroup.js
@@ -2,7 +2,7 @@
 
 import React, { Component } from 'react';
 import CheckboxGroup from './styled/CheckboxGroup';
-import type { ChildrenType } from '../types';
+import type { ChildrenType } from './types';
 
 type Props = {|
   /** Elements to be rendered as a checkbox group. These should be only Checkbox