+++ packages/elements/blanket/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/blanket
 
+## 4.0.8
+- [patch] Packages Flow types for elements components [3111e74](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3111e74)
+
 ## 4.0.7