+++ packages/elements/code/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/code
 
+## 3.0.5
+- [patch] Packages Flow types for elements components [3111e74](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3111e74)
+
 ## 3.0.4
 - [patch] Minor manual bump for packages desync'd from npm [e988c58](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e988c58)