+++ packages/elements/field-text-area/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/field-text-area
 
+## 1.1.3
+- [patch] Packages Flow types for elements components [3111e74](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3111e74)
+
 ## 1.1.2
 - [patch] Change incorrect type info [ce915ea](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ce915ea)